% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC.R
\name{PIC.lm}
\alias{PIC.lm}
\title{PIC method for Linear Model Fits}
\usage{
\method{PIC}{lm}(object, newdata, group_sizes = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from "lm"}

\item{newdata}{An optional data frame in which to look for variables with which to compute PIC. If omitted, the fitted values are used.}

\item{group_sizes}{A scalar or vector indicating the sizes of data partitions. See details.}

\item{...}{further arguments passed to or from other methods.}
}
\value{

}
\description{
PIC method for Linear Model Fits
}
\details{
PIC.lm computes PIC values based on the supplied model.
}
