% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC.R
\name{PIC.lm}
\alias{PIC.lm}
\title{PIC method for Linear Models}
\usage{
\method{PIC}{lm}(object, newdata, group_sizes = NULL, ...)
}
\arguments{
\item{object}{A fitted model object of \code{\link[base]{class}} "lm".}

\item{newdata}{An optional data frame of validation data which may also inform the computation of PIC. If omitted, the training data contained within the fitted model \code{object} are used.}

\item{group_sizes}{A scalar or numeric vector indicating the sizes of \code{newdata} partitions. See 'Details'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
If \code{group_sizes = NULL}, a scalar is returned. Otherwise, \code{newdata} is
returned with an appended column labeled 'PIC' containing group-specific PIC values.
}
\description{
Computation of predictive information criteria for linear models.
}
\details{
PIC.lm computes PIC values based on the supplied model.
}
\seealso{
\code{\link{PIC}}, \code{\link[stats]{lm}}
}
