% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC.R
\name{PIC.mlm}
\alias{PIC.mlm}
\title{PIC method for Multivariable Linear Models}
\usage{
\method{PIC}{mlm}(object, newdata, group_sizes = NULL, bootstraps = NULL, ...)
}
\arguments{
\item{object}{A fitted model object of \code{\link[base]{class}} "mlm".}

\item{newdata}{An optional dataframe of validation data used to compute PIC. If omitted, the training data contained within \code{object} are used. See 'Details'.}

\item{group_sizes}{An optional scalar or numeric vector indicating the sizes of \code{newdata} partitions. If omitted, \code{newdata} is not partitioned. See 'Details'.}

\item{bootstraps}{An optional numeric value indicating the number of bootstrap samples to use for a bootstrapped PIC. See 'Details'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
TBD
}
\description{
Computation of predictive information criteria for multivariable linear models. Currently, computations are
supported for only bivariable linear models.
}
\details{
TBD
}
\seealso{
\code{\link[picR]{PIC}}, \code{\link[stats]{lm}}
}
