% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC.R
\name{PIC}
\alias{PIC}
\title{Compute Predictive Information Criteria (PIC)}
\usage{
PIC(object, ...)
}
\arguments{
\item{object}{a model object to evaluate via predictive information criteria (PIC)}

\item{...}{additional arguments affecting the PIC computed}
}
\value{
The form of the value returned by PIC depends on the class of its argument and the additional arguments supplied. See the documentation of the particular method(s) for details of what is produced by that method.
}
\description{
PIC is a generic function for computing PIC from the results of various model fitting functions. The function invokes particular \emph{methods} which depend on the class of the first argument.
}
\details{
Currently, only a single method exists for this generic function. This method is for the class "lm" of linear models. Additional methods are in development.
}
