% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC.R
\name{PIC}
\alias{PIC}
\title{Predictive Information Criteria}
\usage{
PIC(object, newdata, ...)
}
\arguments{
\item{object}{A fitted model object on which computed PIC value(s) are partially based.}

\item{newdata}{An optional data frame of validation data which may also inform the computation of PIC. If omitted, the training data contained within the fitted model \code{object} are used.}

\item{...}{Further arguments passed to other methods.}
}
\value{
The form of the value returned by \code{PIC} depends on the fitted model class and its method-specific arguments. Details may be found in the documentation of each method.
}
\description{
\code{PIC} is a generic function for computing predictive information criteria.
Depending on the fitted model object \code{\link[base]{class}}, the function invokes the appropriate method for
determining PIC value(s) for the supplied model.
}
\details{
The PIC are criteria developed for model selection within predictive contexts. Given a
set of predictive models, the model with the minimum criterion value is preferred.

The PIC are developed to asymptotically select the candidate model that minimizes the mean
squared error of prediction (MSEP), thus behaving similarly to the the Akaike Information
Criterion (AIC). Contrasting with AIC, however, the PIC do not require the assumption of
validation data that are independent and identically distributed to the set of training data.
The PIC thus offer greater flexibility in predictive model selection applications.
}
\seealso{
\code{\link{PIC.lm}}
}
